[long-running checks] Retry on `ExitError`s only
Use a slightly different approach to long-running check retries.

Before:
Long running checks would always be retried when they returned an
error. This meant that if a check was stopped and returned a non-zero
exit code, it would also be retried. For some reason JMXFetch sometimes
returns a non-zero exit code when killed, so we would end up with
orphaned JMXFetch processes.

After:
The runner doesn't handle long running checks separately. It retries a
check when its `Run` method returns a certain type of error
(`check.RetryableError`). Long running checks are modified to return
that error when they should be retried (right now, only when the
child process returns a non-zero exit code). The jmx check tracks
whether `Stop` was called, and never returns that type of error when
a `Stop` was called.